requirements.txt
pymongo==4.3.3
pandas==1.5.2
psycopg2-binary==2.9.5
regex==2022.10.31
# nltk==3.7
DateTime==4.9
python-dateutil==2.8.2
google-cloud-bigquery==3.4.1
google-auth==2.16.0
db-dtypes==1.0.5
gspread==6.1.2
oauth2client==4.1.3
google-cloud-scheduler==2.13.5